\name{circle.one.point}
\alias{circle.one.point}

\title{
Find a circle around a point
}
\description{
Through simple approximate geometry the perimeter of a circle around a point given in lat and lon is found
}
\usage{
circle.one.point(lat, lon = NULL, rad, n = 10)
}

\arguments{
  \item{lat, lon}{Latitude and longitude in decimal degrees.  }

  \item{rad}{ radius of the circle in \code{nmi}}

  \item{n}{ number of points on the circle perimeter }
}
\details{
possibly later
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
"circle.one.point"
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
