\name{eez}
\alias{eez}
\docType{data}
\title{
EEZ boundaries
}
\description{
A data.frame of EEZ boundaries in the N-Atlanic, taken from some site in Belgium
, a predecessor of \url{http://www.marineregions.org}.}
\usage{data(eez)}
\format{
  The format is:
 chr "eez"
}
\details{
An update from the new site might be a good idea
}
\source{
taken from E. Hj. downloads, should be properly updated from \url{http://www.marineregions.org}. 
}

\examples{
data(eez)
## maybe str(eez) ; plot(eez) ...
}
\keyword{datasets}
