\name{fiveh}
\alias{fiveh}
\docType{data}
\title{
The 500 nmi isobath within the Icelandic EEZ
}
\description{
The main part of the 500 nmi isobath within the Icelandic EEZ, as given
by the list \code{\link[geo]{gbdypi}}
}
\usage{data(fiveh)}
\format{
  The format is:
 chr "fiveh"
}
\details{
Prepared in with a 'manual' script, applying geoinside to remove
unwanted parts of the overall 500 m isobath.
}
\source{
GEBCO
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(fiveh)
## maybe str(fiveh) ; plot(fiveh) ...
}
\keyword{datasets}
