\name{geoextras-package}
\alias{geoextras-package}
\alias{geoextras}
\docType{package}
\title{

  Extras for Hafro's (MRIs) 'geo' package

}
\description{

  Functions and data sets that complement the 'geo' package, much used
at the Marine Research Institute (Hafro), Reykjavik, Iceland. As 'geo'
is not available on 'r-forge', only a few standalone routines will be
added to begin with.
  
}
\details{
\tabular{ll}{
Package: \tab geoextras\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-11-08\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%%~~ the R documentation directory ~~
\keyword{ package }
