\name{r.corners}
\alias{r.corners}

\title{
Find the corner of a statistical rectangle
}

\description{
Finds the corner of a statistical rectangle by adding appropriate offsets
}

\usage{
r.corners(square)
}

\arguments{
  \item{square}{
Statistical rectangle code
}
}

\value{Position of rectangle corner(s), used for text placement.
}

\note{
Experimental, rudimentary, deprecated?
}

\seealso{
\code{\link[geo:deg2rect]{d2r}}, \code{\link[geo:rect2deg]{r2d}}
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
"r.corners"
}

\keyword{arith}
\keyword{manip}
