\name{sap}
\alias{sap}
\docType{data}
\title{
Sea Area Proportion.
}
\description{
Sea area proportion of rectangles abutting onto 
  some parts of the N-Atlantic coastline.
}
\usage{data(sap)}
\format{
  The format is:
 chr "sap"
}

\source{
N-Atlantic coastlines.
}

\seealso{
  \code{\link[geo:rectArea]{drA}}, \code{\link[geo:rectPeri]{drPeri}}.
}

\references{
to be filled in after streamlining with modern packages.
}

\note{
  Other rectangle dimensions might have to be covered, streamlining
  the estimation using modern spatial packages.
}

\examples{
data(sap)
## maybe str(sap) ; plot(sap) ...
}

\keyword{datasets}
