\name{twelvem}
\alias{twelvem}
\docType{data}
\title{Twelve mile zone around Iceland
}
\description{
Twelve mile zone around Iceland as determined by regulations, omitting
Grimsey, data from some official data provider
}
\usage{data(twelvem)}
\format{
  The format is:
 chr "twelvem"
}
\details{
need checking and verification, 
}
\source{
unknown
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(twelvem)
## maybe str(twelvem) ; plot(twelvem) ...
}
\keyword{datasets}
