\name{zbydr}
\alias{zbydr}
\docType{data}
\title{
EEZ division of rectangles.
}
\description{
EEZ division of degree rectangles of dimensions 1 lat and 2 lon.
}
\usage{data(zbydr)}
\format{
  The format is:
 chr "zbydr"
}

\source{
Basic source split of ICES statistical rectangles on EEZs used in blue
whiting negotiations in the 1990s.

}

\seealso{\code{\link{zbyrec}}}

\examples{
data(zbydr)
## maybe str(zbydr) ; plot(zbydr) ...
}
\keyword{datasets}
