\name{zbydr2}
\alias{zbydr2}
\docType{data}
\title{
EEZ divsion of rectangles.
}
\description{
EEZ division of 2 by 4 degree rectangles.
}
\usage{data(zbydr2)}
\format{
  The format is:
 chr "zbydr2"
}

\source{
  Basic source split of ICES statistical rectangles on EEZs used
  in blue whiting negotiations in the 1990s.
}

\seealso{
  \code{link{zbyrec}}
}
    
\examples{
data(zbydr2)
## maybe str(zbydr2) ; plot(zbydr2) ...
}
\keyword{datasets}
