\name{r.corners}
\alias{r.corners}

\title{
Find the corner of a statistical rectangle
}
\description{
Finds the corner of a statistical rectangle by adding appropriate offsets
}
\usage{
r.corners(square)
}

\arguments{
  \item{square}{
Statistical rectangle code
}
}
\details{
unncessary
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link[geo]{d2r}}, \code{\link[geo]{r2d}}
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
"r.corners"
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
